+++ CHANGELOG.md
@@ -13,6 +13,7 @@ colorscad: improved error handling
 - Input file does not exist
 - The wrong 'sort' command is used
 - No colors are found in the input (i.e. due to errors in the input)
+- No file or an empty file is produced by '3mfmerge'
 
 ## [0.3.1] - 2021-07-03
 ### Fixed