+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.4.0] - 2021-07-04
 ### Added
 colorscad:
 - Switch -v to enable verbose logging. Decrease verbosity by default, by using OpenSCAD's --quiet parameter.
@@ -51,7 +53,8 @@ colorscad:
 ### Added
 - Everything! Only supports creating a colored .amf file.
 
-[Unreleased]: https://github.com/jschobben/colorscad/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/jschobben/colorscad/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/jschobben/colorscad/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/jschobben/colorscad/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/jschobben/colorscad/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/jschobben/colorscad/compare/v0.1.0...v0.2.0